Tests/LibWeb/Text/expected/abspos-relayout-inline-containing-block.txt

boundary-before=57.53125,32,120,40
boundary-after=57.53125,32,120,40
content-width=60
partial-layouts=0 full-layouts=1
FAIL
